Many changes in MAKE_MODELS_PARALLEL
Added try-catch structure where "treeqsm" is called. Changed the way
the data is loaded. Now only one point cloud is in the workspace.
Corrected a bug where incomplete QSM was saved as complete QSM. Changed
where the input-structure for each tree is reconstructed. Changed the
coding to separate more the results of the different parallel processes
(less warnings and errors).