Major changes in CYLINDERS
Many comprehensive and small changes: "regions" and "cylinder_fitting"
are combined into "cylinder_fitting” and the process is more adaptive
as it now fits at least 3 (up to 10) cylinders of different lengths for
each region. "lcyl" and "FilRad" parameters are not used anymore.
Surface coverage ("SurfCov") and mean absolute distance ("mad") are
added to the cylinder structure as fields. Surface coverage filtering
is used in the definition of the regions and removing outliers.
“adjustments" has many changes, particularly in the taper corrections
where the parabola-taper curve is fitted to all the data with surface
coverage as a weight. Adjustment of radii based on the parabola is
closer the parabola the smaller the surface coverage. For the stem the
taper correction is the same as for the branches. The minimum and
maximum radii corrections are also modified. Syntax has changed,
particularly for the "cyl"-structure.