Small changes in TREE_SETS
Changed the base height specification. Changed the cylinder fitting
syntax corresponding to the new input and outputs of
"least_squares_cylinder”. Removed "Trunk(Base) = false;" at the
beginning of the "make_tree_connected” as unnecessary and to prevent
errors in a special case where the Trunk is equal to Base. Removed from
the end of “make_tree_connected” the generation of "Trunk" again and
the new call for the function. Increased the minimum distance of a
component to be removed from 8m to 12m.