Major changes in LEAST_SQUARES_CYLINDER
Changed the input parameters of the cylinder and the outputs to the
struct format. Added optional input for weights. Added optional input
"Q", a subset of "P", the cylinder is intended to be fitted in this
subset but it is fitted to "P" to get better estimate of the axis
direction and radius. Added new output, "mad", which is the mean
absolute distance of the points from the surface of the cylinder. Added
new output, "SurfCov", that measures how well the surface of the
cylinder is covered by the points.