Define input for single point cloud in .mat file
Previously, the variable "P" was not defined for lines 89 and beyond if the .mat file with point clouds to be processed only contained a single point cloud. By moving the definition of P outside the nt > 1 condition, P is defined in the case of only one point cloud, and nothing is different if nt > 1.